Add to all relevant places
