+++ src/lib/seam/components/AccessCodeTable/AccessCodeTable.tsx
@@ -78,6 +78,8 @@ export function AccessCodeTable({
   onBack,
   accessCodeFilter = defaultAccessCodeFilter,
   accessCodeComparator = compareByCreatedAtDesc,
+  errorFilter: customErrorFilter = () => true,
+  warningFilter: customWarningFilter = () => true,
   heading = t.accessCodes,
   title = t.accessCodes,
   className,
@@ -134,6 +136,8 @@ export function AccessCodeTable({
     return (
       <NestedEditAccessCodeForm
         accessCodeId={selectedEditAccessCodeId}
+        errorFilter={customErrorFilter}
+        warningFilter={customWarningFilter}
         disableLockUnlock={disableLockUnlock}
         disableCreateAccessCode={disableCreateAccessCode}
         disableEditAccessCode={disableEditAccessCode}
@@ -153,6 +157,8 @@ export function AccessCodeTable({
         onEdit={() => {
           setSelectedEditAccessCodeId(selectedViewAccessCodeId)
         }}
+        errorFilter={customErrorFilter}
+        warningFilter={customWarningFilter}
         disableLockUnlock={disableLockUnlock}
         disableCreateAccessCode={disableCreateAccessCode}
         disableEditAccessCode={disableEditAccessCode}
@@ -169,6 +175,8 @@ export function AccessCodeTable({
     return (
       <NestedCreateAccessCodeForm
         deviceId={deviceId}
+        errorFilter={customErrorFilter}
+        warningFilter={customWarningFilter}
         disableLockUnlock={disableLockUnlock}
         disableCreateAccessCode={disableCreateAccessCode}
         disableEditAccessCode={disableEditAccessCode}
@@ -225,6 +233,8 @@ export function AccessCodeTable({
           accessCodes={filteredAccessCodes}
           onAccessCodeClick={handleAccessCodeClick}
           onAccessCodeEdit={handleAccessCodeEdit}
+          customErrorFilter={customErrorFilter}
+          customWarningFilter={customWarningFilter}
           disableEditAccessCode={disableEditAccessCode}
           disableDeleteAccessCode={disableDeleteAccessCode}
         />
@@ -237,6 +247,8 @@ function Content(props: {
   accessCodes: Array<UseAccessCodesData[number]>
   onAccessCodeClick: (accessCodeId: string) => void
   onAccessCodeEdit: (accessCodeId: string) => void
+  customErrorFilter: (error: AccessCode['errors'][number]) => boolean
+  customWarningFilter: (warning: AccessCode['warnings'][number]) => boolean
   disableEditAccessCode: boolean
   disableDeleteAccessCode: boolean
 }): JSX.Element {
@@ -244,6 +256,8 @@ function Content(props: {
     accessCodes,
     onAccessCodeClick,
     onAccessCodeEdit,
+    customErrorFilter,
+    customWarningFilter,
     disableEditAccessCode,
     disableDeleteAccessCode,
   } = props
@@ -273,6 +287,8 @@ function Content(props: {
         accessCodes={accessCodes}
         filter={filter}
         onFilterSelect={setFilter}
+        errorFilter={customErrorFilter}
+        warningFilter={customWarningFilter}
       />
       {filteredAccessCodes.map((accessCode) => (
         <AccessCodeRow
