+++ src/lib/seam/components/ClimateSettingScheduleTable/ClimateSettingScheduleTable.tsx
@@ -66,6 +66,8 @@ export function ClimateSettingScheduleTable({
   disableDeleteAccessCode = false,
   onBack,
   className,
+  errorFilter: customErrorFilter = () => true,
+  warningFilter: customWarningFilter = () => true,
   disableCreateAccessCode,
   disableEditAccessCode,
 }: ClimateSettingScheduleTableProps): JSX.Element {
@@ -114,6 +116,8 @@ export function ClimateSettingScheduleTable({
     return (
       <NestedClimateSettingScheduleDetails
         climateSettingScheduleId={selectedViewClimateSettingScheduleId}
+        errorFilter={customErrorFilter}
+        warningFilter={customWarningFilter}
         disableLockUnlock={disableLockUnlock}
         disableCreateAccessCode={disableCreateAccessCode}
         disableEditAccessCode={disableEditAccessCode}
